fix: close replyModal & upload cover photo
